Fail reification if there is a destructive update to a term
A limitation of allowing transformations on program terms is that
we may keep the program labels in the expression tree as a way of
maintaining provenance with the CFG representation.

We added the ability to introduce fresh terms, which should be
preferred for most transformations, but there is currently nothing
stopping a user from producing different terms for subexpressions
that share the same label.

This is a (lame) way to enforce that these transformations are
consistent with one another.