Fix use of classes + separate more effectful ops
We should just use the nodes in each class as we maintain them

Also, all calls and loads should not be considered pure operations.

Calls are obviously side-effecting and should not be re-ordered. Loads
for now will be treated conservatively since their ordering is
sensitive to the presence of stores before or after them. Maybe
later we can do an alias analysis to eliminate redundant loads.