Separate intervals analysis from the e-graph framework
They seem to be incompatible, because the representation of nodes in
the e-graph is supposed to carry static and eternal facts about the
program.

With intervals, the facts about a unique node may change depending on
which point in the program we are considering for rewrites.

This becomes a problem when, for example, the analysis tells us that
an SSA variable has a constant value at a particular program point.

The main issue occurs when the node associated with this SSA variable
is unioned with that constant, which then becomes an immutable fact
as far as the e-graph is concerned.

This is a bit disappointing, since we have to separate this out into
a dedicated SCCP pass, but as it turns out the intervals analysis
already does much of the work for us, and this information was not
as useful for rewrite rules as I initially expected.