Temporary fix for searching union nodes
This restores the full `double_xor` rule, which exposed a bug in how
we were searching union nodes.

Some attempts to fix this resulted in a combinatorial blowup on certain
inputs when we try to search all paths.

For now we bias towards the rewritten term, and so the search may miss
more rewrite opportunities on the "original" terms.

However (and interestingly), this doesn't seem to produce a regression
in the current test suite.

Will require more investigation