Local state in contexts + some fixes to evaluation order in `Sm` monad
We can re-use the `Dict` module for maintaining local state in a pure
way during some passes, without resorting to using the monad transformers.

This has to do with how functions like `map` are implemented in the
`Monads` library, which gets us a rather unintuitive evaluation order
when generating fresh variables or labels. There are more places where
we can fix this issue.