Compute instruction dominance by hand
The structure used by manually producing a dominator tree was
inefficient, especially with the LCA computation. It should
be sufficient to just rely on the already-computed dominator
tree for just the basic blocks. We can use the same trick that
gets performed in the `Ssa` pass to check if one definition
appears before another.