Integrate constant phis analysis into the egraph builder
It can benefit from having access to the equivalence classes.
The result is still an over-approximation of the set of unique
values that a block argument can inhabit. This is because our
construction of the e-graph only performs a single pass over
the dominator tree of the function, so we can't currently
revisit loops to discover more equivalences.

As a stop gap measure due to these limitations, we can run the
standalone pass after the e-graph optimizations in hope that it
can pick up some missed opportunities.