Fix issue with patricia trees and pseudo labels
The pseudo labels were starting from a negative number, which caused
issues with the compressed payload-and-branching-bit scheme we use in
the patricia tree implementation (it saves us an extra word of
allocation).

In practice, getting a label that has a large enough number (via the
freshening counter in `Context`) to cause this bug would mean that the
input program is way too large for us to handle anyway (consider
2^57-1 program labels).

This implementation detail means that the patricia tree interface
should be hidden from users and should only be constructed internally
behind safe interfaces.