`imp` arg for ABI
This is a bit of a corner case to deal with variadic calls in SysV
ABI, where RAX needs to hold the number of SSE registers being passed.

It's a problem because in this case RAX would normally be our scratch
register, so the assignment to the register has to be scheduled
immediately before the call.

It's a bit of a hack, but this detail is pretty messy to begin with.