Local isel helper
The reason this is local is because of the `Context.t` type. In the
`Isel.Make` functor, we pass a restricted signature for `Context`,
because we don't want the implementation referring to the machine
of the current context, but we still need the `Machine` passed to
the functor because the type of `Isel.run` depends on it.

It's a bit of a dependency issue. Maybe we'll try to make this nicer
in the future.