Try preprocessing the rules into tables
A majority of the rules will be of the form `move ?x, (op ?x...)`. The
issue with the matching engine is that it would waste a bunch of time
trying to match a node with patterns that obviously wouldn't fit, in
linear order. What we can do instead is just organize this initial
toplevel into tables (similar to the e-graph matching engine), and
just never consider these rules which will never match.