Use the scratch register for handling the overflow
After some thought, it actually seems wrong to punish the user for
this case we refused to handle. To that end, using the scratch
register becomes necessary, so we will exclude it again during
register allocation.

We can't make temporary variables anymore at this stage, because
register allocation has already happened (and running it again could
be expensive). We also can't do this stage earlier, because we don't
know the stack layout until after register allocation, when all
the necessary spills have been performed.