Split assigning stack locations into two phases
Since this is now split, it makes sense to just integrate it into
the register allocation phase.

The idea is that normal stack slots (provided by the user) are
allocated before register allocation, so we end up with a partially
computed stack layout. Any administrative/intermediate computation
needed to make these stack location assigmnents legal can be done
beforehand. Now, the register allocator can better do its job if
we end up needing to create temporaries.

The spilled stack slots are always of a certain form, so that this
treatment is never needed, and thus it can happen after register
allocation.