Fixes bug in LICM optimization
Found this while playing around with the prime numbers example

The algorithm to check loop variance was missing the case where
the loop a node is defined in is a child of the current loop of
the analysis. It is surprising that it took this long for me to
catch this error.