Provide easier API for running isel and regalloc
For added carefulness, we will check that the current context monad
has the same target as the provided machine. This is a deficiency
of the current interface, but I couldn't think of how to do better,
because the `Machine_intf.S` implementation also acts as a witness
for the types that we return/accept.