Fixes a bug related to partial redundancy elimination
Firstly, our implementation of union/merge for patricia trees had a
bug, resulting in duplicate elements in certain cases.

Second, when computing the two closure sets, the union of the sets for
the source (moved-from) nodes should exclude the target node from
the result, because since the `closure` function operates over the
iterated dominance frontier, it is possible that the target node
is reachable from one of its children in the dominator tree.

In all, we were seeing instances where hoisting wasn't happening when
it would be reasonably expected to happen, and the issues were found
to be separate but related.