Fixes e-graph code motion not tracking all possible values
Before, we were only tracking values that were either duplicated or
became unioned, but this led to some values being unnecessarily
rematerialized.

We worked around this with a hack by running the e-graph pass again
with no rewrite rules (plus some extra state passed around), but we'd
rather that this all gets done in a single pass.

With this new change, we'll also optimize some of the data structures
used internally.