Check that the entry block has no parameters
The entry block is "special" in the sense that its parameters are the
same as that of its containing function. We already enforce the
invariant that the entry block shall not be a target of a control-flow
instruction, so it also makes sens that it should never have any
parameters either (because it can never be a join point).